requirements.txt
requests
matplotlib
pandas
python-dotenv
schedule
numpy
tabulate
pyyaml